requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
celery
redis
youtube-transcript-api
openai
finnhub-python
requests
pandas
streamlit
plotly
python-dotenv
apscheduler
feedparser
edgartools
